Move utxos into wallet test util
